リファクタリング
おなじ処理を抽出
